requirements.txt
APScheduler==3.6.0
Django==2.1
django-apscheduler==0.3.0
django-widget-tweaks==1.4.3
dnspython==1.16.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
tzlocal==1.5.1
